integrate with solver
